Add files via upload
add method added. realised that python converts lists to tuples when passed as *args. so I hacked this back to lists. which are easy to append. not sure if this will lead to bad results although the lists are private and need a method to reach them so... maybe I should __the lists.